Move reduction tests into reduction_test.py
